+++ src/polyfill.getMatchedCSSRules.js
@@ -133,7 +133,7 @@
 			selector, score, result = 0;
 
 		while (selector = selectors.shift()) {
-			if (_matchesSelector(element, selector)) {
+			if (element.closest(selector)) {
 				score = calculateScore(selector);
 				result = score > result ? score : result;
 			}
@@ -149,38 +149,6 @@
 		};
 
 		return rules.sort(compareSpecificity);
-	};
-
-	var customMatchesSelector = function (element, selector) {
-		var matches = (element.document || element.ownerDocument).querySelectorAll(selector);
-		var i = 0;
-
-		while (matches[i] && matches[i] !== element) {
-			i++;
-		}
-
-		return matches[i] ? true : false;
-	};
-
-	// Find correct matchesSelector implementation
-	var _matchesSelector = function (element, selector) {
-		var matcher = function (selector) {
-			if (element.matches) {
-				return element.matches(selector);
-			} else if (element.matchesSelector) {
-				return element.matchesSelector(selector);
-			} else if (element.mozMatchesSelector) {
-				return element.mozMatchesSelector(selector);
-			} else if (element.webkitMatchesSelector) {
-				return element.webkitMatchesSelector(selector);
-			} else if (element.msMatchesSelector) {
-				return element.msMatchesSelector(selector);
-			} else {
-				return customMatchesSelector(element, selector);
-			}
-		};
-
-		return matcher(selector);
 	};
 
 	//TODO: not supporting 2nd argument for selecting pseudo elements
@@ -219,7 +187,7 @@
 				}
 
 				// check if this element matches this rule's selector
-				if (_matchesSelector(element, rule.selectorText)) {
+				if (element.closest(rule.selectorText)) {
 					// push the rule to the results set
 					result.push(rule);
 				}